[refactor] background color setting dispatch
